Create TypedExpressionParserFactory and SliceExpression
