+++ parser/expression/TypedExpressionParser.js
@@ -1,12 +1,14 @@
+import { TypedExpressionParserFactory } from '.';
 export class TypedExpressionParser {
   /**
    * Decorates a parser with a typed expression parser.
    * @param {*} parser 
    */
-  constructor (parser) {
-    Object.assign(this, parser || {});
+  constructor () {
+    /* setup functions via factory */
+    let factory = new TypedExpressionParserFactory();
     // special binary operators here
-    this.binaryOperations = {
+    factory.binaryOperations = {
       '*,int,string': 'string',
       '*,string,int': 'string',
       '+,string,string': 'string',
@@ -16,25 +18,24 @@ export class TypedExpressionParser {
     let logical = ['or', 'nor', 'xor', 'and', 'nand',
       '<', '>', '<=', '>=', '=', '!=', '&', '|', '^'];
     for (let op of logical) {
-      this.binaryOperations[[op, 'flag', 'flag']] =
-      this.binaryOperations[[op, 'flag', 'int']] =
-      this.binaryOperations[[op, 'int', 'flag']] =
-      this.binaryOperations[[op, 'int', 'int']] = 'flag';
+      factory.binaryOperations[[op, 'flag', 'flag']] =
+      factory.binaryOperations[[op, 'flag', 'int']] =
+      factory.binaryOperations[[op, 'int', 'flag']] =
+      factory.binaryOperations[[op, 'int', 'int']] = 'flag';
     }
     // relational operators always return flag
     let relational = ['<', '>', '<=', '>=', '=', '!='];
     for (let op of relational) {
-      this.binaryOperations[[op, 'flag', 'flag']] =
-      this.binaryOperations[[op, 'int', 'int']] =
-      this.binaryOperations[[op, 'string', 'string']] = 'flag';
+      factory.binaryOperations[[op, 'flag', 'flag']] =
+      factory.binaryOperations[[op, 'int', 'int']] =
+      factory.binaryOperations[[op, 'string', 'string']] = 'flag';
     }
     // arithmetic operators return int
     let arithmetic = ['&', '|', '^', '<<', '>>', '+', '-', '*', '/'];
     for (let op of arithmetic) {
-      this.binaryOperations[[op, 'int', 'int']] = 'int';
+      factory.binaryOperations[[op, 'int', 'int']] = 'int';
     }
-    // unary operations here
-    this.unaryOperations = {
+    factory.unaryOperations = {
       '+,int': 'int',
       '-,int': 'int',
       '~,int': 'int',
@@ -64,59 +65,50 @@ export class TypedExpressionParser {
     let expression = () => {
       let result = this.typedParenthesisExpression() || this.typedData();
       if (result === undefined) return undefined;
-      let checkpoint = this.index;
       this.whitespace();
       // check for post-group operators
-      if (this.peek() == '(') {
-        // perform embed operation
-        if (result[1] === 'string') {
-          let args = this.argumentListExpression();
-          if (args !== null) {
-            return {
-              _id: 'embed',
-              text: result[0],
-              args: args
-            };
-          }
-        } else {
-          this.error('no definition for <' + result[1] + '> (...)');
-        }
-      } else if (this.character('[')) {
-        // slice
-        if (result[1] === 'string') {
-          let slice = this.sliceExpression();
-          if (slice !== null && this.character(']')) {
-            return [{
-              _id: 'op',
-              operator: '[]',
-              operands: [result[0], slice]
-            }, 'string'];
-          }
-        }
-        // TODO: add slice for non-string in the future
-      } else if (this.character('<')) {
-        // type cast
-        let type = this.dataType();
-        if (type !== null && this.character('>')) {
+      let isString = result[1] === 'string';
+      if (isString) {
+        let args = this.argumentListExpression();
+        if (args !== undefined) {
           return [{
-            _id: 'op',
-            operator: '<>',
-            operands: [type, result[0]]
-          }, type];
+            _id: 'embed',
+            text: result[0],
+            args: args
+          }, 'string'];
         }
       }
-      this.undo(this.index - checkpoint);
+      if (isString) {
+        let slice = this.sliceExpression();
+        if (slice !== undefined) {
+          return [{
+            _id: 'op',
+            operator: '[]',
+            operands: [result[0], slice]
+          }, 'string'];
+        }
+      }
+      let cast = this.typeCast();
+      if (cast !== undefined) {
+        return [{
+          _id: 'op',
+          operator: '<>',
+          operands: [cast, result[0]]
+        }, cast];
+      }
       return result;
     };
     // incrementally wrap expression parser depending on order of operations
     for (let unaryOperations of orderOfUnaryOperations) {
-      expression = this.createTypedUnaryOperation(expression, unaryOperations);
+      expression = factory.createTypedUnaryOperation(expression,
+        unaryOperations);
     }
     for (let binaryOperations of orderOfBinaryOperations) {
-      expression = this.createTypedLeftToRightBinaryOperation(expression,
+      expression = factory.createTypedLeftToRightBinaryOperation(expression,
         binaryOperations);
     }
     // finally, add this to class methods
+
     /**
      * Checks if the next token is an expandable expression.
      */
@@ -162,127 +154,5 @@ export class TypedExpressionParser {
       }
     }
     return this.undo(this.index - checkpoint);
-  }
-
-  /**
-   * Gets the return type of a binary operation.
-   * @param {*} op    the name of a binary operator
-   * @param {*} type1 the data type at the left-hand side operand
-   * @param {*} type2 the data type at the right-hand side operand
-   */
-  binaryOperationReturnType (op, type1, type2) {
-    if (type1 === undefined || type2 === undefined) {
-      return undefined;
-    }
-    return this.binaryOperations[[op, type1, type2]];
-  }
-
-  /**
-   * Gets the return type of a unary operation.
-   * @param {*} op    the name of the unary operation
-   * @param {*} type1 the data type of the operand
-   */
-  unaryOperationReturnType (op, type1) {
-    if (type1 === undefined) {
-      return undefined;
-    }
-    return this.unaryOperations[[op, type1]];
-  }
-
-  /**
-   * Creates a pre-unary operation that operates in the correct evaluation
-   * order.
-   * @param {*} nextLevel a higher priority operation
-   * @param {*} operands  a list of unary operands of the same priority
-   */
-  createTypedUnaryOperation (nextLevel, operands) {
-    let self = this;
-    return currentLevel;
-    function currentLevel () {
-      let checkpoint = self.index;
-      for (let operand of operands) {
-        let name = operand.trim();
-        if (operand[0] !== ' ' || self.whitespace()) {
-          if (self.phrase(name)) {
-            if (operand[operand.length - 1] !== ' ' || self.whitespace()) {
-              // at this point, unary operation has already been parsed
-              // try again for nested unary operations of the same level
-              let S = currentLevel();
-              if (S !== undefined) {
-                let type = self.unaryOperationReturnType(name, S[1]);
-                // warn if unknown return type
-                if (type === undefined && S[1] !== undefined) {
-                  self.warning('undefined unary operation: ' +
-                    name + ' <' + S[1] + '>');
-                }
-                // perform unary operation
-                return [{
-                  _id: 'op',
-                  operator: name,
-                  operands: [S[0]]
-                }, type];
-              }
-            }
-          }
-        }
-        self.undo(self.index - checkpoint);
-      }
-      return nextLevel();
-    }
-  }
-
-  /**
-   * Creates a parser for a left-to-right binary operation that operates in
-   * the correct evaluation order.
-   * @param {*} nextLevel a higher priority operation
-   * @param {*} operands  a list of binary operands of the same priority
-   */
-  createTypedLeftToRightBinaryOperation (nextLevel, operands) {
-    let self = this;
-    return () => {
-      let checkpoint = self.index;
-      let S1 = nextLevel();
-      if (S1 !== undefined) {
-        let hasOperator = true;
-        while (hasOperator) {
-          hasOperator = false;
-          let operatorCheckpoint = self.index;
-          // try one of the operands
-          for (let operand of operands) {
-            let name = operand.trim();
-            if (operand[0] !== ' ' || self.whitespace()) {
-              if (self.phrase(name)) {
-                if (operand[operand.length - 1] !== ' ' || self.whitespace()) {
-                  let S2 = nextLevel();
-                  if (S2 !== undefined) {
-                    let type = self.binaryOperationReturnType(name, S1[1],
-                      S2[1]);
-                    // collate binary operation, left-to-right
-                    S1 = [{
-                      _id: 'op',
-                      operator: name,
-                      operands: [S1[0], S2[0]]
-                    }, type];
-                    hasOperator = true;
-                    // warn if return type is undefined
-                    if (type === undefined &&
-                      S1[1] !== undefined &&
-                      S2[1] !== undefined) {
-                      self.warning('undefined binary operation: <' + S1[1] +
-                        '> ' + operand + ' ' + name + ' <' + S2[1] + '>');
-                    }
-                    // break out of operator loop, continue parsing level
-                    break;
-                  }
-                }
-              }
-            }
-            self.undo(self.index - operatorCheckpoint);
-          }
-        }
-        return S1;
-      }
-      return self.undo(self.index - checkpoint);
-    };
   }
 }
